Add finished bump-formula workflow
This workflow is triggered by a "repository dispatch" event. A repository
dispatch event in this repository can be triggered by making a POST
request to https://api.github.com/repos/weibeld/homebrew-core/dispatches.

The "type" of the repository dispatch event must be "bump-formula" (this
is an arbitrary name), and its payload must contain the fields
"formula", "url", and "sha256":

  - formula: the formula to bump
  - url:     the URL of the new distribution for the formula
  - sha256:  the SHA256 of the distribution archive

Note that the current and new version of the formula are auto-detected
from the provided URL by Homebrew.

Here's a complete example request for triggering this workflow:

curl -X POST \
  -H "Authorization: Bearer <TOKEN>" \
  -H "Accept: application/vnd.github.v3+json" \
  https://api.github.com/repos/weibeld/homebrew-core/dispatches \
  -d '
    {
      "event_type": "bump-formula",
      "client_payload": {
        "formula": "k1s",
        "url": "https://github.com/weibeld/k1s/releases/download/0.3.0/k1s-0.3.0.tar.gz",
        "sha256": "447bd602c1d486a47ebedd88f2c3f4c872ed68139bf47cdff1b2347c5d1c5313"
      }
    }'